Unix: Store pointer values in a more portable way

+++ src/include/cpu_emulation.h
@@ -21,6 +21,9 @@
 #define CPU_EMULATION_H
 
 
+#include "sheeptypes.h"
+
+ 
 /*
  *  Memory system
  */
@@ -49,11 +52,11 @@ struct KernelData {
 };
 
 // RAM and ROM pointers (allocated and set by main_*.cpp)
-extern uint32 RAMBase;			// Base address of Mac RAM
+extern uintptr RAMBase;			// Base address of Mac RAM
 extern uint32 RAMSize;			// Size address of Mac RAM
 extern uint8 *RAMBaseHost;		// Base address of Mac RAM (host address space)
 
-extern uint32 ROMBase;			// Base address of Mac ROM
+extern uintptr ROMBase;			// Base address of Mac ROM
 extern uint8 *ROMBaseHost;		// Base address of Mac ROM (host address space)
 
 // Mac memory access functions
